+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Fixed
+- LinkedDataBox#hasTagId() and  LinkedDataBox#hasTag() were broken with a missing predicate.
 
 ## 0.0.4 - 2016-01-18
 ### Added